build user and device relation, and user to device authorization

+++ server/socket/deviceSocket.js
@@ -2,6 +2,8 @@
 const path = require("path");
 const fs = require("fs");
 const Device = require("../models/Device.js");
+const User = require("../models/User.js");
+
 // const Peripheral = require('../models/Peripheral.js')
 const espSetup = require("../esp32/espSetup");
 
@@ -15,35 +17,45 @@ deviceSocket = (socket) => {
     console.log("this is plist : ", pList);
     ackCallBack(pList);
   });
+
   socket.on("addDevice", async (data, ackCallBack) => {
     try {
+      // Optional: You can also use socket.user._id directly if JWT middleware includes it
+      const userEmail = socket.user?.email;
+      if (!userEmail) throw new Error("User not authenticated");
+
+      const user = await User.findOne({ email: userEmail });
+      if (!user) throw new Error("User not found");
+
       const peripherals = {};
       data.peripherals.forEach((p) => {
         peripherals[p] = `value of ${p}`;
       });
-      console.log(data.picture);
-      console.log(data.pictureName);
+
       const doc = new Device({
         name: data.name,
         location: data.location,
         dictVariables: peripherals,
+        user: user._id, // associate device with the authenticated user
+        image: data.picture || undefined, // optionally include image if provided
       });
+
       const device = await doc.save();
-      ackCallBack(device.id);
+      ackCallBack(device._id);
     } catch (err) {
       ackCallBack(-1);
-      // socket.emit("errorSetup", err.message);
-      console.log(err);
+      console.error("Error in addDevice:", err);
     }
   });
 
   socket.on("setupDevice", async (deviceId) => {
-    const device = await Device.findOne({ id: deviceId });
+    await checkDeviceOwnership(socket, deviceId);
+    const device = await Device.findById(deviceId);
     console.log("this is device : ", device);
     try {
       const pList = Array.from(device.dictVariables.keys());
       console.log("this is pList : ", pList);
-      await espSetup(deviceId, pList, socket);
+      await espSetup(device.id, pList, socket);
     } catch (err) {
       await Device.findByIdAndDelete(device._id);
       socket.emit("errorSetup", {
@@ -59,7 +71,8 @@ deviceSocket = (socket) => {
 
   socket.on("deleteDevice", async (deviceId, ackCallBack) => {
     try {
-      const device = await Device.findOneAndDelete({ id: deviceId });
+      await checkDeviceOwnership(socket, deviceId);
+      const device = await Device.findByIdAndDelete(deviceId);
 
       ackCallBack(`the device: ${device._id} is deleted successfully`);
     } catch (err) {
@@ -67,6 +80,23 @@ deviceSocket = (socket) => {
       ackCallBack(`error in deleting the device, ${err.message}`);
     }
   });
+};
+
+const checkDeviceOwnership = async (socket, deviceId) => {
+  if (!socket.user || !socket.user.id) {
+    throw new Error("User not authenticated");
+  }
+
+  const device = await Device.findById(deviceId);
+  if (!device) {
+    throw new Error("Device not found");
+  }
+
+  if (device.user.toString() !== socket.user.id.toString()) {
+    throw new Error("Unauthorized: You do not own this device");
+  }
+
+  return device; // optionally return device for reuse
 };
 
 module.exports = deviceSocket;
